refactor: useKeywardSearch 경로 수정 및 useCallback 수정
